2.0.3 Ajustes

+++ src/components/Programa.js
@@ -12,8 +12,13 @@ const Programa = ({ programa, onUpdate, onDelete }) => {
     setIsEditable(!isEditable);
   };
 
-  const handleDeleteClick = () => {
-    onDelete(programa.id); 
+  const handleDeleteOrCancelClick = () => {
+    if (isEditable) {
+      setIsEditable(false);
+      setFormState(programa || {});
+    } else if (onDelete) {
+      onDelete(programa.id);
+    }
   };
 
   const handleChange = (e) => {
@@ -31,12 +36,10 @@ const Programa = ({ programa, onUpdate, onDelete }) => {
       boxShadow: '0 2px 10px rgba(0, 0, 0, 0.5)',
       padding: '20px',
       borderRadius: '8px',
-      //border: '2px solid #5eb219',
       width: '30%',
       minWidth: '300px',
       margin: '10px auto',
     }}>
-
       <div style={{
         display: 'flex',
         flexDirection: 'column',
@@ -51,15 +54,11 @@ const Programa = ({ programa, onUpdate, onDelete }) => {
         }}>Nombre</label>
         <TextField
           disabled={!isEditable}
-          sx={{
-            width: '100%',
-          }}
+          sx={{ width: '100%' }}
           name="nombre"
           value={formState.nombre || ''}
           variant="outlined"
-          InputProps={{
-            readOnly: !isEditable,
-          }}
+          InputProps={{ readOnly: !isEditable }}
           onChange={handleChange}
         />
       </div>
@@ -78,15 +77,11 @@ const Programa = ({ programa, onUpdate, onDelete }) => {
         }}>Nombre Corto</label>
         <TextField
           disabled={!isEditable}
-          sx={{
-            width: '100%',
-          }}
+          sx={{ width: '100%' }}
           name="nombreCorto"
           value={formState.nombreCorto || ''}
           variant="outlined"
-          InputProps={{
-            readOnly: !isEditable,
-          }}
+          InputProps={{ readOnly: !isEditable }}
           onChange={handleChange}
         />
       </div>
@@ -101,23 +96,24 @@ const Programa = ({ programa, onUpdate, onDelete }) => {
         <Button sx={{
           backgroundColor: '#5eb219',
           color: '#fff',
-          '&:hover': {
-            backgroundColor: '#4cae14',
-          },
+          '&:hover': { backgroundColor: '#4cae14' },
           marginRight: '10px',
         }} onClick={handleEditClick}>
           {isEditable ? 'Guardar' : 'Editar'}
         </Button>
         <Button sx={{
-          backgroundColor: '#b2195e',
+          backgroundColor: isEditable ? '#d81b60' : '#b2195e',
           color: '#fff',
           '&:hover': {
-            backgroundColor: '#d81b60',
+            backgroundColor: isEditable ? '#4cae14' : '#d81b60',
           },
-        }} onClick={handleDeleteClick}>Borrar</Button>
+        }} onClick={handleDeleteOrCancelClick}>
+          {isEditable ? 'Cancelar' : 'Borrar'}
+        </Button>
       </div>
     </div>
   );
 };
 
 export default Programa;
+
